backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-multipart==0.0.12
pyjwt==2.9.0
python-dotenv==1.0.1
pydantic==2.9.2
pillow==10.4.0
